new Beginning!

+++ src/util/Vector2.java
@@ -1,10 +1,10 @@
 package util;
 
 public class Vector2 {
-    private double x;
-    private double y;
+    private int x;
+    private int y;
 
-    public Vector2(double x, double y) {
+    public Vector2(int x, int y) {
         this.x = x;
         this.y = y;
     }
@@ -15,12 +15,12 @@ public class Vector2 {
         return new Vector2(this.x, this.y);
     }
 
-    public Vector2 addX(double x) {
+    public Vector2 addX(int x) {
         this.x += x;
         return new Vector2(this.x, this.y);
     }
 
-    public Vector2 addY(double y) {
+    public Vector2 addY(int y) {
         this.y += y;
         return new Vector2(this.x, this.y);
     }
@@ -36,10 +36,10 @@ public class Vector2 {
         return false;
     }
 
-    public double x() {
+    public int x() {
         return this.x;
     }
-    public double y() {
+    public int y() {
         return this.y;
     }
